# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/auth/login": {
      "post": {
        "operationId": "AccountLoginController_request",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MagicLinkRequestDto"
              }
            }
          }
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MagicLinkAcceptedDto"
                }
              }
            }
          }
        },
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/auth/login/preview": {
      "post": {
        "operationId": "AccountLoginController_preview",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginTokenDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginPreviewDto"
                }
              }
            }
          }
        },
        "tags": [
          "account"
        ]
      }
    },
    "/api/v1/auth/login/verify": {
      "post": {
        "operationId": "AccountLoginController_verify",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginVerifyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LoginCompletedDto"
                }
              }
            }
          }
        },
        "tags": [
          "account"
        ]
      }
    },
...
    "/api/v1/spaces/{id}/memberships": {
      "get": {
        "operationId": "AccessController_list",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MembershipListDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/spaces/{id}/memberships/{actorId}/role": {
      "post": {
        "operationId": "AccessController_assign",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AssignRoleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "access"
        ]
      }
    },
    "/api/v1/spaces/{id}/memberships/{actorId}/remove": {
      "post": {
        "operationId": "AccessController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "actorId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PolicyChangedDto"
                }
              }
            }
          }
        },
        "security": [
          {
            "fc_sid": []
          }
        ],
        "tags": [
          "access"
        ]
      }
    },
...
  },
  "info": {
    "title": "Lịch Gia Đình API",
    "description": "",
    "version": "1",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "fc_sid"
      }
    },
    "schemas": {
...
      "MagicLinkRequestDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string"
          }
        },
        "required": [
          "email"
        ]
      },
      "MagicLinkAcceptedDto": {
        "type": "object",
        "properties": {
          "accepted": {
            "type": "boolean",
            "enum": [
              true
            ]
          }
        },
        "required": [
          "accepted"
        ]
      },
      "LoginTokenDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          }
        },
        "required": [
          "token"
        ]
      },
      "LoginActorDto": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "label"
        ]
      },
      "LoginPreviewDto": {
        "type": "object",
        "properties": {
          "actors": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/LoginActorDto"
            }
          }
        },
        "required": [
          "actors"
        ]
      },
      "LoginVerifyDto": {
        "type": "object",
        "properties": {
          "token": {
            "type": "string"
          },
          "actor_id": {
            "type": "string"
          },
          "device_id": {
            "type": "string"
          },
          "label": {
            "type": "string"
          }
        },
        "required": [
          "token",
          "actor_id",
          "device_id"
        ]
      },
      "LoginCompletedDto": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "device_id": {
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "device_id"
        ]
      },
...
      "ManagedMembershipDto": {
        "type": "object",
        "properties": {
          "actor_id": {
            "type": "string"
          },
          "role_id": {
            "type": "string"
          },
          "role_key": {
            "type": "string"
          },
          "display_name": {
            "type": "string"
          }
        },
        "required": [
          "actor_id",
          "role_id",
          "role_key",
          "display_name"
        ]
      },
      "ManagedRoleDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "role_key": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "matrix": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        },
        "required": [
          "id",
          "role_key",
          "name",
          "matrix"
        ]
      },
      "MembershipListDto": {
        "type": "object",
        "properties": {
          "memberships": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManagedMembershipDto"
            }
          },
          "roles": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ManagedRoleDto"
            }
          },
          "can_manage": {
            "type": "boolean"
          }
        },
        "required": [
          "memberships",
          "roles",
          "can_manage"
        ]
      },
      "AssignRoleDto": {
        "type": "object",
        "properties": {
          "role_id": {
            "type": "string"
          }
        },
        "required": [
          "role_id"
        ]
      },
      "PolicyChangedDto": {
        "type": "object",
        "properties": {
          "policy_version": {
            "type": "string"
          }
        },
        "required": [
          "policy_version"
        ]
      },
...
    }
  }
}